Performance enhancements.
itkSphereConvolutionFilter now assumes the input PSF image is a radial
profile if its size in the y-direction is 1 and performs the sampling
appropriately. Also, added a performance tweak that prevents redundant
computation of whether a sample point is within the boundary of an image.

itkGibsonLanniBSFImageSource now no longer performs a rotational extrusion
on the PSF profile, leaving this to itkSphereConvolutionFilter.